To Version 1.4

+++ Client/PokerResult.cpp
@@ -19,7 +19,7 @@ CString CPokerResult::GetPokerString(byte* p, int count)
 
 	// Sort value
 	std::list<byte> ltVal;
-	for (size_t i = 0; i < count; i++)
+	for (int i = 0; i < count; i++)
 	{
 		ltVal.push_back(p[i] % 12);
 	}
